+++ packages/cordova/src/android/AdMob.java
@@ -6,6 +6,8 @@ import android.os.Bundle;
 import android.provider.Settings;
 import android.util.Log;
 import java.math.BigDecimal;
+import java.util.ArrayList;
+
 import org.apache.cordova.CallbackContext;
 import org.apache.cordova.CordovaInterface;
 import org.apache.cordova.CordovaPlugin;
@@ -29,6 +31,8 @@ public class AdMob extends CordovaPlugin {
 
     private static final String TEST_APPLICATION_ID = "ca-app-pub-3940256099942544~3347511713";
 
+    private ArrayList<PluginResult> waitingForReadyCallbackContextResults = new ArrayList<PluginResult>();
+
     @Override
     public void initialize(CordovaInterface cordova, CordovaWebView webView) {
         super.initialize(cordova, webView);
@@ -41,8 +45,14 @@ public class AdMob extends CordovaPlugin {
     public boolean execute(String actionKey, JSONArray args, CallbackContext callbackContext) {
         Action action = new Action(args);
         if (Actions.READY.equals(actionKey)) {
+            if (waitingForReadyCallbackContextResults == null) {
+                return false;
+            }
             readyCallbackContext = callbackContext;
-
+            for (PluginResult result : waitingForReadyCallbackContextResults) {
+              readyCallbackContext.sendPluginResult(result);
+            }
+            waitingForReadyCallbackContextResults = null;
             JSONObject data = new JSONObject();
             try {
                 data.put("platform", "android");
@@ -57,6 +67,8 @@ public class AdMob extends CordovaPlugin {
             return BannerAd.executeHideAction(action, callbackContext);
         } else if (Actions.BANNER_SHOW.equals(actionKey)) {
             return BannerAd.executeShowAction(action, callbackContext);
+        } else if (Actions.INTERSTITIAL_IS_LOADED.equals(actionKey)) {
+            return InterstitialAd.executeIsLoadedAction(action, callbackContext);
         } else if (Actions.INTERSTITIAL_LOAD.equals(actionKey)) {
             return InterstitialAd.executeLoadAction(action, callbackContext);
         } else if (Actions.INTERSTITIAL_SHOW.equals(actionKey)) {
@@ -106,7 +118,11 @@ public class AdMob extends CordovaPlugin {
 
         PluginResult result = new PluginResult(PluginResult.Status.OK, event);
         result.setKeepCallback(true);
-        readyCallbackContext.sendPluginResult(result);
+        if (readyCallbackContext == null) {
+          waitingForReadyCallbackContextResults.add(result);
+        } else {
+          readyCallbackContext.sendPluginResult(result);
+        }
     }
 
     private String getApplicationID() {