+++ packages/cordova/src/ios/AMSInterstitial.swift
@@ -5,6 +5,10 @@ class AMSInterstitial: AMSAdBase, GADInterstitialDelegate {
         interstitial = nil
     }
 
+    func isLoaded() -> Bool {
+        return (interstitial?.isReady == true)
+    }
+
     func load(request: GADRequest) {
         let interstitial = GADInterstitial(adUnitID: adUnitID)
         self.interstitial = interstitial
@@ -32,7 +36,6 @@ class AMSInterstitial: AMSAdBase, GADInterstitialDelegate {
     @objc
     func interstitialDidDismissScreen(_ adInterstitial: GADInterstitial) {
         plugin.emit(eventType: AMSEvents.interstitialClose)
-        fitAds()
     }
 
     @objc