+++ packages/core/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.4](https://github.com/admob-plus/admob-plus/compare/@admob-plus/core@0.7.3...@admob-plus/core@0.7.4) (2019-04-28)
+
+**Note:** Version bump only for package @admob-plus/core
+
+
+
+
+
+## [0.7.3](https://github.com/admob-plus/admob-plus/compare/@admob-plus/core@0.7.2...@admob-plus/core@0.7.3) (2019-04-26)
+
+**Note:** Version bump only for package @admob-plus/core
+
+
+
+
+
+## [0.7.2](https://github.com/admob-plus/admob-plus/compare/@admob-plus/core@0.7.0...@admob-plus/core@0.7.2) (2019-04-11)
+
+**Note:** Version bump only for package @admob-plus/core
+
+
+
+
+
+## [0.7.1](https://github.com/admob-plus/admob-plus/compare/@admob-plus/core@0.7.0...@admob-plus/core@0.7.1) (2019-04-11)
+
+**Note:** Version bump only for package @admob-plus/core
+
+
+
+
+
+# [0.7.0](https://github.com/admob-plus/admob-plus/compare/@admob-plus/core@0.6.1...@admob-plus/core@0.7.0) (2019-03-18)
+
+
+### Features
+
+* **cordova:** add interstitial.isLoaded method ([f52a624](https://github.com/admob-plus/admob-plus/commit/f52a624))
+
+
+
+
+
 ## [0.6.1](https://github.com/admob-plus/admob-plus/compare/@admob-plus/core@0.6.0...@admob-plus/core@0.6.1) (2018-12-11)
 
 **Note:** Version bump only for package @admob-plus/core